Selenium WebDriver: 1. Worked on the code review comments. All comments are addressed. 2. Restructure the whole project to follow the standard maven archetype 3. Deleted the obsolete files from previous tasks: the files of application, unit tests and smoke and regression suites.
