+++ src/main/java/com/coherentsolutions/aqa/web/makarevich/constants/TimeOutConstants.java
@@ -1,0 +1,9 @@
+package com.coherentsolutions.aqa.web.makarevich.constants;
+
+import java.time.Duration;
+
+public class TimeOutConstants {
+    public static final Duration SHORT_TIMEOUT = Duration.ofSeconds(5);
+    public static final Duration MEDIUM_TIMEOUT = Duration.ofSeconds(10);
+    public static final Duration LONG_TIMEOUT = Duration.ofSeconds(30);
+}
