+++ src/main/java/com/coherentsolutions/aqa/web/makarevich/constants/UrlConstants.java
@@ -1,0 +1,15 @@
+package com.coherentsolutions.aqa.web.makarevich.constants;
+
+import com.coherentsolutions.aqa.web.makarevich.utils.UrlUtils;
+
+/**
+ List of URL constants
+ */
+public class UrlConstants {
+    public static final String YANDEX_MAIL_URL = UrlUtils.getProperty("yandexmail.url");
+    public static final String MULTISELECT_TEST_URL = UrlUtils.getProperty("multiselectedtest.url");
+    public static final String ALERTS_TEST_URL = UrlUtils.getProperty("alertstest.url");
+    public static final String LOADING_DATA_DYNAMICALLY_URL = UrlUtils.getProperty("loadingdatadynamically.url");
+    public static final String PROGRESS_BAR_URL = UrlUtils.getProperty("progressbar.url");
+    public static final String TABLE_TEST_URL = UrlUtils.getProperty("tabletest.url");
+}
