+++ src/main/java/com/coherentsolutions/aqa/web/makarevich/utils/WebDriverUtils.java
@@ -1,0 +1,21 @@
+package com.coherentsolutions.aqa.web.makarevich.utils;
+
+import org.openqa.selenium.WebDriver;
+import org.openqa.selenium.chrome.ChromeDriver;
+import org.openqa.selenium.firefox.FirefoxDriver;
+
+public class WebDriverUtils {
+    public static WebDriver setWebDriver() {
+        WebDriver driver;
+        switch (ConfigUtils.getProperty("driver.browser")) {
+            case "chrome":
+                driver = new ChromeDriver();
+                break;
+            default:
+                driver = new FirefoxDriver();
+                break;
+        }
+        driver.manage().window().maximize();
+        return driver;
+    }
+}
