--- src/main/java/parser/NoSuchFileException.java
@@ -1,12 +1,0 @@
-package parser;
-
-public class NoSuchFileException extends RuntimeException {
-
-    public NoSuchFileException() {
-
-    }
-
-    public NoSuchFileException(String message, Throwable t) {
-        super(message, t);
-    }
-}
