--- src/main/java/parser/Parser.java
@@ -1,11 +1,0 @@
-package parser;
-
-import shop.Cart;
-
-import java.io.File;
-
-public interface Parser {
-
-    void writeToFile(Cart cart);
-    Cart readFromFile(File file);
-}
