--- src/main/resources/eugen-cart.json
@@ -1,1 +1,0 @@
-{"cartName":"eugen-cart","realItems":[{"weight":1400.0,"name":"BMW","price":22103.9}],"virtualItems":[{"sizeOnDisk":8500.0,"name":"Microsoft office","price":30.0}],"total":26560.68}
