--- src/test/java/constants/TimeOutConstants.java
@@ -1,9 +1,0 @@
-package constants;
-
-import java.time.Duration;
-
-public class TimeOutConstants {
-    public static final Duration SHORT_TIMEOUT = Duration.ofSeconds(5);
-    public static final Duration MEDIUM_TIMEOUT = Duration.ofSeconds(10);
-    public static final Duration LONG_TIMEOUT = Duration.ofSeconds(30);
-}
