--- src/test/java/constants/UrlConstants.java
@@ -1,14 +1,0 @@
-package constants;
-
-import utils.ConfigUtils;
-import utils.UrlUtils;
-
-public class UrlConstants {
-    public static final String RESOURCE_PATH = ConfigUtils.getProperty("resource.path");
-    public static final String YANDEX_MAIL_URL = UrlUtils.getProperty("yandexmail.url");
-    public static final String MULTISELECT_TEST_URL = UrlUtils.getProperty("multiselectedtest.url");
-    public static final String ALERTS_TEST_URL = UrlUtils.getProperty("alertstest.url");
-    public static final String LOADING_DATA_DYNAMICALLY_URL = UrlUtils.getProperty("loadingdatadynamically.url");
-    public static final String PROGRESS_BAR_URL = UrlUtils.getProperty("progressbar.url");
-    public static final String TABLE_TEST_URL = UrlUtils.getProperty("tabletest.url");
-}
