--- src/test/java/testng/VirtualItemTest.java
@@ -1,33 +1,0 @@
-package testng;
-
-import org.testng.annotations.AfterClass;
-import org.testng.annotations.BeforeClass;
-import org.testng.annotations.Test;
-import org.testng.asserts.SoftAssert;
-import shop.VirtualItem;
-
-
-public class VirtualItemTest {
-    private VirtualItem virtualItem;
-
-    @BeforeClass(alwaysRun = true)
-    void setUp() {
-        virtualItem = new VirtualItem();
-    }
-
-    @Test(groups = "Smoke")
-    void testVirtualItem() {
-        double sizeOnDisk = 500.25;
-        virtualItem.setSizeOnDisk(sizeOnDisk);
-
-        SoftAssert softAssert = new SoftAssert();
-        softAssert.assertEquals(500.25, virtualItem.getSizeOnDisk());
-        softAssert.assertEquals("Class: class shop.VirtualItem; Name: null; Price: 0.0; Size on disk: 500.25", virtualItem.toString());
-        softAssert.assertAll();
-    }
-
-    @AfterClass(alwaysRun = true)
-    void tearDown() {
-        virtualItem = null;
-    }
-}
