--- src/test/resources/suits/alltest-testng.xml
@@ -1,10 +1,0 @@
-<!DOCTYPE suite SYSTEM "https://testng.org/testng-1.0.dtd" >
-
-<suite name="All tests">
-    <test name="All tests">
-        <packages>
-            <package name="testng"/>
-            <package name="selenium"/>
-        </packages>
-    </test>
-</suite>
