--- src/test/resources/suits/regression-testng.xml
@@ -1,14 +1,0 @@
-<!DOCTYPE suite SYSTEM "https://testng.org/testng-1.0.dtd" >
-
-<suite name="Regression">
-    <test name="Regression">
-        <groups>
-            <run>
-                <exclude name="Smoke"/>
-            </run>
-        </groups>
-        <packages>
-            <package name="testng"/>
-        </packages>
-    </test>
-</suite>
