--- src/test/resources/suits/smoke-testng.xml
@@ -1,14 +1,0 @@
-<!DOCTYPE suite SYSTEM "https://testng.org/testng-1.0.dtd" >
-
-<suite name="Smoke">
-    <test name="SmokeTest">
-        <groups>
-            <run>
-                <include name="Smoke"/>
-            </run>
-        </groups>
-        <packages>
-            <package name="testng"/>
-        </packages>
-    </test>
-</suite>
