+++ suites/all.xml
@@ -1,0 +1,9 @@
+<!DOCTYPE suite SYSTEM "https://testng.org/testng-1.0.dtd" >
+
+<suite name="All tests">
+    <test name="All tests">
+        <packages>
+            <package name="com.coherentsolutions.aqa.java.web.makarevich.*"/>
+        </packages>
+    </test>
+</suite>
